+++ storage/spanstore/spanstoremetrics/decorator.go
@@ -2,7 +2,7 @@
 // Copyright (c) 2017 Uber Technologies, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-package metrics
+package spanstoremetrics
 
 import (
 	"context"
