+++ storage/spanstore/spanstoremetrics/package_test.go
@@ -1,7 +1,7 @@
 // Copyright (c) 2023 The Jaeger Authors.
 // SPDX-License-Identifier: Apache-2.0
 
-package metrics
+package spanstoremetrics
 
 import (
 	"testing"
